requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
jinja2
python-multipart
pydantic
passlib[bcrypt]